Possible enhancements to state reporting
I noticed that in the parse section you were capturing a very long endsWith section, and it included the 4 character network address specific to this device on your network. Only the last 4 is needed to determine on/off.

I also added an attribute reporting request in the configure section. If this switch supports it, it would send a message back to the hub everytime the on/off is changed regardless of the method. You would need to live log to check for those messages and their format, but hopefully one or both of these changes helps capture physical state changes at the switch. (You could check the thread "The state of Zigbee with attribute reporting" for more info on the command I added to configure. Hope it helps!